feat: move subscription details into util file
